flask-app/app/requirements.txt
azure-core==1.32.0
azure-servicebus==7.13.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
